feat(#11): 메시지 라우팅 구조 정의
